+++ net-deleter/src/handlers/clear_buffer_handler.rs
@@ -1,5 +1,7 @@
 use std::error::Error;
 use std::sync::Arc;
+use net_core_api::api::primitives::none::None;
+use net_core_api::core::encoder_api::Encoder;
 use net_deleter_api::api::buffer::ClearBufferRequestDTO;
 use net_component::handler::network_service_handler::NetworkServiceHandler;
 use net_core_api::api::envelope::envelope::Envelope;
@@ -41,7 +43,7 @@ impl NetworkServiceHandler for ClearBufferHandler {
             return Err(DeleteError::DbError(deletable_data_type, err).into());
         }
         let _ = transaction.commit().await;
-        Ok(Envelope::new(tenant_id, "none", b""))
+        Ok(Envelope::new(tenant_id, None::get_data_type(), &None::default().encode()))
     }
 
     fn get_handler_type(&self) -> String {