+++ net-updater/src/component/updater_component.rs
@@ -1,6 +1,7 @@
 use std::net::SocketAddr;
 use std::sync::Arc;
 
+use component_core::materialized_view::core::common::get_common_materialized_view_manager;
 use net_component::component::network_service_component::NetworkServiceComponent;
 use net_component::handler::network_service_handler_manager::NetworkServiceHandlerManager;
 use net_component::handler::network_service_handler_manager_builder::NetworkServiceHandlerManagerBuilder;
@@ -9,6 +10,7 @@ use sqlx::Postgres;
 
 use crate::config::Config;
 use crate::handlers::refreshers::refresh_packets_handler::RefreshPcapParsedDataHandler;
+use crate::handlers::refreshers::materialized_views_refresh_handler::MaterializedViewsRefreshHandler;
 use crate::handlers::updaters::packets_network_id_handler::UpdatePacketsNetworkIdHandler;
 use component_core::connection_pool;
 
@@ -40,6 +42,7 @@ impl UpdaterComponent {
             NetworkServiceHandlerManagerBuilder::default()
                 .add_handler(UpdatePacketsNetworkIdHandler::default().boxed())
                 .add_handler(RefreshPcapParsedDataHandler::default().boxed())
+                .add_handler(MaterializedViewsRefreshHandler::new(Box::new(get_common_materialized_view_manager())).boxed())
                 .build()
         )
     }